feat: fetch file on file preview open (in case file data is previewable, e.g. text files)
